refactor(allocator): move to top level
move allocator to top level dir since it's used only in messagequeue now
