+++ allocator/allocator_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/ipfs/go-graphsync/responsemanager/allocator"
+	"github.com/ipfs/go-graphsync/allocator"
 	"github.com/ipfs/go-graphsync/testutil"
 )
